Add function playRound
Add pseudocode to describe how the function is intended
to work.
Based on that created the actual function which compares the
choices of the user and the computer. Depending on their choices
the function decides who has won and show the result to the user.
The function counts the wins of both players and stores them in
variables.